minor error

+++ network/request.go
@@ -190,7 +190,7 @@ func (reqst *Request) ExecuteRequest() {
 		reqst.w.WriteHeader(200)
 
 		if reqst.Active == nil {
-			reqst.Active = active.NewActive(maxbufsize,reqst, map[string]interface{}{"DBMS": db.DBMSManager, "Parameters": func() *parameters.Parameters {
+			reqst.Active = active.NewActive(int64(maxbufsize),reqst, map[string]interface{}{"DBMS": db.DBMSManager, "Parameters": func() *parameters.Parameters {
 				return reqst.Parameters()
 			}, "DBQuery": func(alias string, query string, args ...interface{}) (dbquery *db.DBQuery) {
 				dbquery = reqst.DbQuery(alias, query, args)
